added unit test

+++ src/main/java/com/getir/readingisgood/service/AuthService.java
@@ -1,6 +1,8 @@
 package com.getir.readingisgood.service;
 
+import com.getir.readingisgood.exceptions.ResourceNotFoundException;
 import com.getir.readingisgood.model.response.GenericReturnValue;
+import com.getir.readingisgood.utils.ErrorMessages;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
@@ -16,6 +18,9 @@ public class AuthService {
 
     public GenericReturnValue<Boolean> logout() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null) {
+            throw new ResourceNotFoundException(ErrorMessages.AUTHENTICATION_NOT_FOUND);
+        }
         userTokenService.deleteToken(authentication.getName());
         log.info("Logout username: {}", authentication.getName());
         return new GenericReturnValue<>(true);
